Adapt workflow sections to `aiida-core==1.0.0`
The majority of the main section and the appendix on the advanced
workchain has been adapted. The scripts that are referenced in the text
that used to be hosted in the `aiida-demos` repository are moved here.
This is to simplify keeping the scripts and excerpts in the text in
sync.

Images and snippets that are included in the text are now put in the
`include` folder directly in the page directory. Since relative imports
are now possible, since we switched from Jekyll to Sphinx, it makes more
sense to keep them there than in a top level `assets` directory.