Rewrite the appendix on restarting calculations
The old way used some utility scripts written specifically for the
`aiida-quantumespresso` plugin. However, these are QE specific solutions.
The new `get_builder_restart` makes restarting so easy that the utility
functions are a bit superfluous and the native implementation works for
all plugins. Therefore we rewrite the appendix to use this method, which
will also make it clearer how it can be used for other use cases by not
hiding the code in a magic utility function.