Automating jupyter notebook generation
The script to generate both versions (tutorials and
solutions) of the tutorial is integrated with the
Makefile. Moreover I put back in place the correct
version both for the 2018 and the 2019 tutorial.
This means that the jupyter notebook in aiida_demos
(in the branch tutorial_2019) will be now deleted and
the one provided here is the "official" version.

I now display the tutorial version of the notebook,
and I provide the solutions as a downloadable file.
Further changes can be done in this repository,
to the *-template.ipynb file.