Updated appendix on queries (stylistic) (#30)
Converted also 2018 tutorial
Improved 2019 tutorial page

Note that for some reason in-line math does not work for me...
Things like
```
:math:`a^2`
```
don't work, I don't know why.

I also changed to SVG format (vectorial) rather than PNG - we
might even think to swap to mathjax that maybe it's even better?

A comment for references to figures: note that the whole documentation
is a single big document, so when duplicating figures between tutorials,
one need to distinguish the labels. In this case, I added a `2018`
in the label of the figure of the 2018 tutorial.