Add a download link for `perovskites` export archive (#35)
The query builder part of the tutorial depends on some data being
present in the database. This data would be imported for the virtual
machine so it would be available to users during the tutorial. Importing
the archive was not enough, as it relied on some extras being set, which
at the time were not exportable. Therefore a script had to be run after
importing that would set these extras and recreated the groups for the
pseudo potential families. The archive was not made available publicly
which means that users running the tutorial in their own environment and
database could not perform this last section.

These problems have now been addressed:

 * The archive was migrated from v0.3 to v0.4 with `aiida-core==1.0.0b3`
 * The migrated archive was imported into a clean database
 * The pseudo potential family groups were recreated
 * The extras were set on the calculations and structure nodes:
   - Structure formula was set on the calculation node
   - `A` and `B` element of `ABO3` perovskite formule set on structure
 * The 3 pseudo family groups and 3 calculation groups were exported.
   - Only direct inputs and outputs of the calculations were included

The text in the query builder notebook has been updated to include a
link to this archive. For now this is a dropbox link, but should be
replaced with a permanent link once the archive is fully tested and
confirmed to be correct.