setup and appendix 1 update
conversion of appendix 1 from md to rst, and some restructuring,
as it was a bit unclear in my opinion.

in the setup section, i refer to aiida documentation 1.0b1 (as putting latest
might get the wrong documentation later on), but needs to be changed if there is a
later installation.
I assumed the slack page and VM address will be OK before the release of the tutorial