Section 2: second run-through
Added a new section near the beginning where the user is instructed to
import an export archive, that will be made available at a permanent
public URL later. This will contain the calculations on perovskites
using Quantum ESPRESSO that will be used later in the query builder
section. In this section, this data will also serve to illustrate the
`verdi process` commands.

I have rewritten quite a bit of text concerning the language around and
the explanation of identifiers. Since identifiers play such a big role
in `verdi` and this section is about how to use the command line, I
thought it merited some more detail.