Section 6: second run-through (#61)
A big part of this was missed in the first run-through apparently. Many
of the advanced scripts, such as the equation of state work chain and the
pressure convergence work chain used the old API and were in general bad
examples of keeping provenance. I have significantly rewritten them,
such that they should now run with `aiida-core==1.0.0` and the
provenance should be kept a lot better by using calculation and work
functions were applicable.