Fix the querybuilder notebook instructions of section 5 (#80)
* Fix the querybuilder notebook instructions of section 5

The instructions now tell the user to download the notebook, for which a
link is provided, and to save it to their machine. They can then launch
the jupyter server and open the notebook.

A section was added to the very first section to explain how these
download links can be used straight on a machine to which they are SSH
connected using `wget`. The text in section 5 has been adapted to this
to remove explicit mentioning of the virtual machine, in order to not
confuse those that are following it without our VM.

Note that I also removed the `.gitignore` in the `notebooks` folder. It
is true that there is a script to generate the two versions from the
template, however, this was necessary when they were still hosted in the
`aiida-demos` directory. For the tutorial VM this was then checked out
and as preparational step, the `scripts/make_notebook.py` was ran.
However, now these versions need to always be present in the repository
for them to be downloadable. They should occassionally just be
regenerated through the script if the content in the template changes.

* PR suggestions by @ltalirz