Section 2 and 4: setup profile, computer and codes (#85)
This section will explain the concept of profiles by creating one
specifically for the tutorial through `verdi quicksetup`. This will also
make the tutorial usable for people that are not running through the
virtual machine but their own installation.

Additionally, section 4 will now give instructions on how to setup a
computer and a code. For people running on the virtual machine, prepared
yaml configuration files that can be used with the `--config` option are
provided. People running on their own installation will have to adapt
thse. Instructions to do so are added in an appendix which is currently
just a stub.