👌 IMPROVE: Final section in workchain module
Large overhaul of the final section in the "Work Chain" module, which shows
how to submit calculation jobs inside a work chain.

Switches to using the `AddWorkChain` as a starting point for adding the
calculation job, now that the added multiply step is added in an exercise.
This has the added benefit that the example is smaller, and they can switch to
using the calculation job in the `MultiplyAddWorkChain` as an exercise.

The explanation of some steps is also more extensive, and the steps for
launching the work chain are split up a bit.

Finally, some more exercises are added at the end of the section.