👌IMPROVE: Workflows - work functions module
Massive overhaul of the work functions section, adding an extensive section on
calculation functions, and some of the more subtle challenges that have been
left out of previous tutorials.

Adds an example based on the `rescale` method. This shows the challenges
regarding converting nodes when adapting a Python function for AiiDA.

Also adds several exercises, also in the work function part.