📝 DOCS: Updates for compatibility with AiiDA v2.5.1
These minor changes fix issues and deprecation warnings that arose when going through the tutorial with the latest AiiDA v2.5.1. The modifications are mainly:
- Update the `verdi code setup` to `verdi code create` commands
- Update the SSSP version from 1.1 to 1.3 to reflect the change of default in `aiida-pseudo`
- Prepend `core` in the `DataFactory` calls where applicable, as well as `query.queryhelp` -> `query.as_dict()` in the `querying` notebook

One could also update the outputs, for instance, to reflect the change from `Info` to `Report` for the `verdi process` commands, as well as adding the additional information on caching which is now shown to avoid confusion, though it's not really necessary for being able to work through the material.

I understand that this page was initially intended for the AiiDA courses, so it's good to have a frozen version for each edition of the course. Though, as we'd like to eventually integrate this material into the main documentation (see issue #XXXX), I think it's good to have it updated now.

Lastly, a few deprecation warnings remain, for example:
```shell
- aiida/cmdline/groups/dynamic.py:145: AiidaDeprecationWarning: Relying on `_get_cli_options` is deprecated. The options should be defined through a `pydantic.BaseModel` that should be assigned to the `Config` class attribute. (this will be removed in v3)
- aiida/orm/nodes/data/code/legacy.py:42: AiidaDeprecationWarning: The `Code` class is deprecated. To create an instance, use the `aiida.orm.nodes.data.code.installed.InstalledCode` or `aiida.orm.nodes.data.code.portable.PortableCode` for a "remote" or "local" code, respectively. If you are using this class to compare type, e.g. in `isinstance`, use `aiida.orm.nodes.data.code.abstract.AbstractCode`. (this will be removed in v3)
- aiida/orm/nodes/data/code/legacy.py:172: AiidaDeprecationWarning: `Code.get_description` method is deprecated, use the `description` property instead. (this will be removed in v3)
```
however, I think these have to be adressed in the `src`.